Extract shared RL server models and build_model_config into utils.
Move JobConfig, request DTOs, and build_model_config out of http_server.py
and ray_server.py so both transports share one definition.
